Added serial recieve code
Added serial recieve code, changed some file names, and removed some extraneous files.
