--- Blink_LED/project.properties
@@ -1,2 +1,0 @@
-name=Blink_LED
-#assetOtaDir=assets
