+++ serialrecieve/project.properties
@@ -1,0 +1,2 @@
+name=serialtransmit
+#assetOtaDir=assets
