+++ serialtransmit/project.properties
@@ -1,0 +1,2 @@
+name=Blink_LED
+#assetOtaDir=assets
